feat: refined layout strategy, add main feed route
