+++ Server/Feed/Sources/App/configure.swift
@@ -1,11 +1,5 @@
 import Vapor
 
-// configures your application
 public func configure(_ app: Application) throws {
-    // uncomment to serve files from /Public folder
-    // app.middleware.use(FileMiddleware(publicDirectory: app.directory.publicDirectory))
-
-    // register routes
-    
     try routes(app)
 }
