+++ Server/Feed/Sources/App/routes.swift
@@ -1,11 +1,6 @@
 import Vapor
 
 func routes(_ app: Application) throws {
-    app.get { req in
-        return "It works!"
-    }
-
-    app.get("hello") { req -> String in
-        return "Hello, world!"
-    }
+    let feedController = FeedController()
+    try app.register(collection: feedController)
 }
